Rewrite generateDictionary as generateDictionaryEntries
